backend/requirements.txt
fastapi>=0.123.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
werkzeug>=3.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
